refactored Repository once more to support keyword 'base' more work on inheritance

+++ Projects/Core/Core/Interpreter.cpp
@@ -94,7 +94,7 @@ void Interpreter::expression(Object *result, TokenIterator& start)
 void Interpreter::garbageCollector()
 {
 	for ( Symbols::reverse_iterator it = mSymbols.rbegin(); it != mSymbols.rend(); ) {
-		if ( it->first != KEYWORD_THIS &&
+		if ( it->first != KEYWORD_BASE && it->first != KEYWORD_THIS &&
 			 it->second && it->second->getType() == Symbol::IType::ObjectSymbol ) {
 			mRepository->removeReference(static_cast<Object*>(it->second));
 		}
@@ -598,10 +598,10 @@ void Interpreter::process_identifier(TokenIterator& token, Object *result, Token
     }
 
 	Object *symbol = static_cast<Object*>(resolve(identifier));
-	if ( !symbol ) {
+	if ( !symbol ) {	// we tried to access an unknown symbol
 		throw Utils::Exceptions::UnknownIdentifer("identifier '" + identifier + "' not found", token->position());
 	}
-	if ( symbol->isConst() ) {	// we tried to modify a const symbol (i.e. member, parameter or local variable)
+	if ( symbol->isConst() ) {	// we tried to modify a const symbol (i.e. member, parameter or constant local variable)
 		throw Utils::Exceptions::ConstCorrectnessViolated("tried to modify const symbol '" + identifier + "'", token->position());
 	}
 	if ( symbol->isMember() && this->isConst() ) {	// we tried to modify a member in a const method
