+++ Projects/Core/Core/Method.cpp
@@ -213,7 +213,7 @@ ControlFlow::E Method::execute(const ParameterList& params, Object *result)
 void Method::garbageCollector()
 {
 	for ( Symbols::reverse_iterator it = mSymbols.rbegin(); it != mSymbols.rend(); ) {
-		if ( it->first != KEYWORD_THIS &&
+		if ( it->first != KEYWORD_BASE && it->first != KEYWORD_THIS &&
 			 it->second && it->second->getType() == Symbol::IType::ObjectSymbol ) {
 			mRepository->removeReference(static_cast<Object*>(it->second));
 		}
